+++ beeref/fileio/export.py
@@ -216,6 +216,7 @@ class SceneToSVGExporter(SceneExporterBase):
                 pixmap, imgformat = item.pixmap_to_bytes(
                     apply_grayscale=True,
                     apply_contrast=True,
+                    apply_lineart=True,
                     apply_crop=True)
                 pixmap = base64.b64encode(pixmap).decode('ascii')
                 element = ET.Element(